TimeScaleを追加。 ゲームオーバー、クリア時にスローになる演出を追加。 アリのステートの問題は、酸攻撃ステート時に共通処理を書き忘れていたから（修正済み）。

+++ DX11MyEngine/Source/Header/EffectManager.h
@@ -25,7 +25,7 @@ private:
     ::EffekseerRenderer::RendererRef m_EfkRenderer; // エフェクトの描画に必要なもの
     std::unordered_map<std::string,Effekseer::EffectRef> m_EffectMap; // エフェクトのパラメータを持つ
 
-    int32_t m_Timer;
+    float m_Timer;
 
 public:
     EffectManager();
